% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study.R
\name{case_study}
\alias{case_study}
\title{Render the case study}
\usage{
case_study(case_file = "CDC-bio-test1.Rmd")
}
\arguments{
\item{case_file}{File to render, by default "CDC-bio-test1.Rmd"}
}
\value{
html-document
}
\description{
Render the case study
}
